refactor(web): split web/routes/pipeline.py by concern (#481 item 3)
pipeline.py had grown to 3,067 lines — 3x the next-largest routes
module — holding CRUD, triage, long-tail, search-plan, and the
/api/_index self-documentation machinery all in one file. Mechanical
AST-carve onto the #496 RouteRegistration registry: search-plan,
triage, long-tail, and api-index handlers move to their own modules
(web/routes/search_plan.py, triage.py, long_tail.py, api_index.py)
together with their ROUTES declarations, wired into web/server.py's
ALL_ROUTES merge. pipeline.py drops to ~2,170 lines (core CRUD,
dashboard/decisions, import jobs, beets-distance).

No behavior change — a throwaway AST-identity gate proved 0 added /
0 removed / 0 changed function-or-class bodies (after import-strip)
and 0 count mismatches across all 36 ROUTES entries between the old
monolith and the new file set. Tests mirror 1:1: triage and long-tail
contract tests split out of tests/web/test_routes_pipeline.py into
tests/web/test_routes_triage.py / test_routes_long_tail.py; the
already-separate search-plan test file renamed to match its module.
