+++ tests/web/test_routes_pipeline.py
@@ -1,8 +1,10 @@
 #!/usr/bin/env python3
-"""Contract tests for web/routes/pipeline.py read routes, triage, long-tail, beets-distance.
+"""Contract tests for web/routes/pipeline.py read routes + beets-distance.
 
 Split from tests/test_web_server.py (#408). Shared harness in
-tests/web/_harness.py.
+tests/web/_harness.py. Triage and long-tail contract tests moved to
+tests/web/test_routes_triage.py / tests/web/test_routes_long_tail.py
+(#481 item 3), following web/routes/pipeline.py's own split.
 """
 
 from datetime import datetime, timezone
@@ -1493,408 +1495,6 @@ class TestBeetsDistanceRouteContract(_FakeDbWebServerCase):
         status, _ = self._get("/api/beets-distance/100/2048516")
         self.assertEqual(status, 404)
 
-
-class TestTriageRouteContracts(_FakeDbWebServerCase):
-    """U17 contracts for ``GET /api/triage/<id>`` and ``GET /api/triage/list``.
-
-    Both endpoints wrap ``lib.triage_service`` (U15) — the same service
-    layer ``pipeline-cli triage show/list`` (U16) wraps. The wire shape
-    on the cohort + composition payloads is the
-    ``msgspec.to_builtins(TriageResult)`` shape verbatim, so the same
-    Struct round-trips through ``msgspec.convert`` on both sides (CLI
-    ⇄ API surface symmetry).
-
-    Tests drive the real ``compose_triage_for_request`` and
-    ``list_triage`` paths against the per-test :class:`FakePipelineDB`
-    (``self.db``) — no service-layer mocking,
-    per ``code-quality.md`` § MOCKS: LEAF-SEAM ONLY. Seeded rows use
-    production-shape values: ``datetime.datetime`` for timestamps via
-    ``make_request_row``'s defaults, real ``FieldResolutionRow`` /
-    ``SearchLogRow`` via the typed seed helpers.
-    """
-
-    # The frontend triage drawer renders these top-level fields out of
-    # ``msgspec.to_builtins(TriageResult)``. Pin every one so a future
-    # field rename can't silently break the JS without flipping a test.
-    SHOW_REQUIRED_FIELDS = {
-        "request_meta", "unfindable", "field_quality", "search_forensics",
-    }
-
-    # ``request_meta`` fields the frontend depends on for the "Artist –
-    # Album (year) #N" header + identity probes (failure_class, source,
-    # search_filetype_override).
-    SHOW_REQUEST_META_FIELDS = {
-        "id", "artist_name", "album_title", "year", "status", "source",
-        "mb_release_id", "discogs_release_id", "release_group_year",
-        "is_va_compilation", "catalog_number", "failure_class",
-        "search_filetype_override",
-    }
-
-    LIST_REQUIRED_FIELDS = {"results", "next_after", "page_size", "filter"}
-
-    # --- /api/triage/<id> -------------------------------------------------
-
-    def test_show_returns_200_with_required_fields_and_roundtrips(self):
-        """Happy path: a seeded request composes through to a 200 with
-        the full TriageResult shape, and the response body round-trips
-        through ``msgspec.convert(payload, type=TriageResult)`` — the
-        wire-boundary contract per CLI ⇄ API symmetry."""
-        from lib.triage_service import TriageResult
-        self.db.seed_request(make_request_row(
-            id=4242,
-            artist_name="Triage Artist",
-            album_title="Triage Album",
-            status="wanted",
-            failure_class="search_not_converting",
-            unfindable_category="artist_absent",
-            unfindable_categorised_at=datetime(2026, 5, 20, tzinfo=timezone.utc),
-            last_artist_probe_at=datetime(2026, 5, 22, tzinfo=timezone.utc),
-            last_artist_probe_match_count=0,
-        ))
-
-        status, data = self._get("/api/triage/4242")
-
-        self.assertEqual(status, 200)
-        _assert_required_fields(self, data, self.SHOW_REQUIRED_FIELDS,
-                                "triage show response")
-        _assert_required_fields(self, data["request_meta"],
-                                self.SHOW_REQUEST_META_FIELDS,
-                                "triage show request_meta")
-        # The wire shape is exactly the Struct shape — round-trip proves
-        # no field drift / coercion happened at the boundary.
-        composed = msgspec.convert(data, type=TriageResult)
-        self.assertEqual(composed.request_meta.id, 4242)
-        self.assertEqual(composed.request_meta.artist_name, "Triage Artist")
-        self.assertEqual(
-            composed.request_meta.failure_class, "search_not_converting",
-        )
-        # Unfindable struct populated because the seeded row has signals.
-        self.assertIsNotNone(composed.unfindable)
-        assert composed.unfindable is not None
-        self.assertEqual(composed.unfindable.category, "artist_absent")
-
-    def test_show_returns_404_when_request_id_missing(self):
-        """Unknown request id → 404 with ``error`` + ``request_id`` in body
-        so the frontend can surface "not found" with the right id."""
-        status, data = self._get("/api/triage/99999")
-        self.assertEqual(status, 404)
-        self.assertIn("error", data)
-        self.assertEqual(data["request_id"], 99999)
-
-    def test_show_returns_404_for_non_int_path(self):
-        """A non-numeric path segment doesn't even match the regex
-        (which requires ``\\d+``), so the route table itself replies
-        404. This test pins the route-table contract (no silent
-        coercion to a different handler)."""
-        status, _ = self._get("/api/triage/not-an-int")
-        # The regex r"^/api/triage/(\d+)$" does not match — falls
-        # through to the catch-all 404.
-        self.assertEqual(status, 404)
-
-    # --- /api/triage/list --------------------------------------------------
-
-    def test_list_filter_unfindable_returns_200_with_required_fields(self):
-        """A seeded unfindable request shows up under
-        ``filter=unfindable`` with the documented envelope shape."""
-        from lib.triage_service import TriageResult
-        self.db.seed_request(make_request_row(
-            id=10, artist_name="Stuck Artist",
-            unfindable_category="artist_absent",
-            unfindable_categorised_at=datetime(2026, 5, 20, tzinfo=timezone.utc),
-        ))
-        # Decoy row without any unfindable signal — must NOT appear in
-        # the filtered cohort.
-        self.db.seed_request(make_request_row(
-            id=11, artist_name="Healthy Artist", status="imported",
-        ))
-
-        status, data = self._get("/api/triage/list?filter=unfindable")
-
-        self.assertEqual(status, 200)
-        _assert_required_fields(self, data, self.LIST_REQUIRED_FIELDS,
-                                "triage list response")
-        self.assertEqual(data["filter"], "unfindable")
-        self.assertEqual(data["page_size"], 50)
-        # Only the unfindable row should be returned.
-        self.assertEqual(len(data["results"]), 1)
-        composed = msgspec.convert(data["results"][0], type=TriageResult)
-        self.assertEqual(composed.request_meta.id, 10)
-        # Page is shorter than page_size → next_after is None
-        # (cohort exhausted).
-        self.assertIsNone(data["next_after"])
-
-    def test_list_filter_data_quality_status_filters_by_status_column(self):
-        """``filter=data_quality:status=<status>`` (issue #374 canonical
-        form) returns only requests with at least one
-        ``album_request_field_resolutions`` row whose ``status`` column
-        matches the spec. Mirrors what
-        ``lib/field_resolver_service.py::_classify_lookup_exception``
-        actually writes."""
-        # Seeded request A: has a release_group_year resolution in the
-        # sticky 4xx-client bucket — matches.
-        self.db.seed_request(make_request_row(id=20))
-        self.db.record_field_resolution(
-            request_id=20, field_name="release_group_year",
-            status="unresolved_4xx_client", reason_code="http_400",
-        )
-        # Seeded request B: has a field resolution but a different
-        # status bucket — must NOT appear.
-        self.db.seed_request(make_request_row(id=21))
-        self.db.record_field_resolution(
-            request_id=21, field_name="catalog_number",
-            status="unresolved_mirror_unavailable",
-            reason_code="ConnectionError",
-        )
-
-        status, data = self._get(
-            "/api/triage/list?filter=data_quality:status=unresolved_4xx_client"
-        )
-
-        self.assertEqual(status, 200)
-        self.assertEqual(
-            data["filter"], "data_quality:status=unresolved_4xx_client",
-        )
-        self.assertEqual(len(data["results"]), 1)
-        self.assertEqual(data["results"][0]["request_meta"]["id"], 20)
-
-    def test_list_filter_data_quality_reason_filters_by_reason_code(self):
-        """``filter=data_quality:reason=<code>`` filters on the
-        ``reason_code`` column (HTTP code specifier — http_400,
-        http_410, http_422, etc.)."""
-        # Seeded request A: 4xx-client status, reason_code=http_400 — matches.
-        self.db.seed_request(make_request_row(id=22))
-        self.db.record_field_resolution(
-            request_id=22, field_name="release_group_year",
-            status="unresolved_4xx_client", reason_code="http_400",
-        )
-        # Seeded request B: 4xx-client status but reason_code=http_410 — excluded.
-        self.db.seed_request(make_request_row(id=23))
-        self.db.record_field_resolution(
-            request_id=23, field_name="catalog_number",
-            status="unresolved_4xx_client", reason_code="http_410",
-        )
-
-        status, data = self._get(
-            "/api/triage/list?filter=data_quality:reason=http_400"
-        )
-
-        self.assertEqual(status, 200)
-        self.assertEqual(data["filter"], "data_quality:reason=http_400")
-        self.assertEqual(len(data["results"]), 1)
-        self.assertEqual(data["results"][0]["request_meta"]["id"], 22)
-
-    def test_list_invalid_filter_returns_400_with_valid_filters_array(self):
-        """An unparseable filter spec surfaces as a 400 carrying
-        ``error`` + a ``valid_filters`` array, so the operator can
-        self-correct without leaving the network response."""
-        status, data = self._get("/api/triage/list?filter=garbage_value")
-
-        self.assertEqual(status, 400)
-        self.assertIn("error", data)
-        self.assertIn("valid_filters", data)
-        self.assertIsInstance(data["valid_filters"], list)
-        # The four canonical scalar forms must be advertised.
-        self.assertIn("all", data["valid_filters"])
-        self.assertIn("unfindable", data["valid_filters"])
-        self.assertIn("data_quality", data["valid_filters"])
-        self.assertIn("search_not_converting", data["valid_filters"])
-
-    def test_list_limit_caps_results_and_emits_next_after_cursor(self):
-        """When the page is exactly ``limit`` long the response carries
-        ``next_after`` = last request_id so the operator can paginate."""
-        for rid in (30, 31, 32):
-            self.db.seed_request(make_request_row(
-                id=rid, status="imported",
-            ))
-
-        status, data = self._get("/api/triage/list?filter=all&limit=2")
-
-        self.assertEqual(status, 200)
-        self.assertEqual(data["page_size"], 2)
-        self.assertEqual(len(data["results"]), 2)
-        # Page was full → next_after is the last id in the page.
-        self.assertEqual(data["next_after"],
-                         data["results"][-1]["request_meta"]["id"])
-
-    def test_list_default_filter_when_query_string_omitted(self):
-        """Missing ``filter=`` defaults to ``all`` so a bare hit on
-        ``/api/triage/list`` is meaningful."""
-        self.db.seed_request(make_request_row(id=40, status="imported"))
-
-        status, data = self._get("/api/triage/list")
-
-        self.assertEqual(status, 200)
-        self.assertEqual(data["filter"], "all")
-        self.assertGreaterEqual(len(data["results"]), 1)
-
-    def test_list_rejects_non_int_limit(self):
-        status, data = self._get("/api/triage/list?filter=all&limit=abc")
-        self.assertEqual(status, 400)
-        self.assertIn("error", data)
-
-    def test_list_rejects_out_of_bounds_limit(self):
-        status, data = self._get("/api/triage/list?filter=all&limit=500")
-        self.assertEqual(status, 400)
-        self.assertIn("error", data)
-
-    def test_list_rejects_non_int_after(self):
-        status, data = self._get(
-            "/api/triage/list?filter=all&after=not-an-int")
-        self.assertEqual(status, 400)
-        self.assertIn("error", data)
-
-
-class TestLongTailRouteContracts(_FakeDbWebServerCase):
-    """U1 contract for ``GET /api/pipeline/long-tail``.
-
-    Wraps ``lib.long_tail_service.list_long_tail`` — the same service
-    ``pipeline-cli long-tail`` wraps (CLI ⇄ API symmetry). Drives the
-    real service + DB cohort query against a fresh :class:`FakePipelineDB`
-    (no service mocking, per MOCKS: LEAF-SEAM ONLY). Banding's beets
-    collaborators (``check_beets_library`` / ``_beets_db`` /
-    ``compute_library_rank``) are the leaf seam — patched at
-    ``web.server`` only when a test exercises an in-library band.
-    """
-
-    # The frontend long-tail list renders these fields per row out of the
-    # serialized ``LongTailRow``. Pin every one so a rename can't silently
-    # break the JS.
-    ROW_REQUIRED_FIELDS = {
-        "id", "artist_name", "album_title", "year", "status", "source",
-        "mb_release_id", "discogs_release_id", "target_format",
-        "min_bitrate", "search_filetype_override", "unfindable_category",
-        "band", "in_flight_rescue",
-        # Card meta (year · MB/Discogs · N tracks) + on-disk spectral strip.
-        "track_count", "current_spectral_grade", "current_spectral_bitrate",
-        # The accept-sibling control + siblings panel read the rg straight
-        # off the row — the single-row refetch must not drop it (#398).
-        "mb_release_group_id",
-    }
-    ENVELOPE_REQUIRED_FIELDS = {"results", "band", "count"}
-
-    def test_missing_row_bands_missing_and_imported_absent(self):
-        """AE1 at the HTTP boundary: a wanted row with no beets album
-        bands ``missing``; an imported request is absent from the
-        result. (No beets configured → everything Missing.)"""
-        from lib.long_tail_service import LongTailRow
-        self.db.seed_request(make_request_row(
-            id=1, status="wanted", mb_release_id="rel-1",
-            artist_name="Vanishing", album_title="Lost"))
-        self.db.seed_request(make_request_row(
-            id=2, status="imported", mb_release_id="rel-2"))
-
-        status, data = self._get("/api/pipeline/long-tail")
-
-        self.assertEqual(status, 200)
-        _assert_required_fields(self, data, self.ENVELOPE_REQUIRED_FIELDS,
-                                "long-tail envelope")
-        self.assertEqual(data["count"], 1)
-        self.assertIsNone(data["band"])
-        row = data["results"][0]
-        _assert_required_fields(self, row, self.ROW_REQUIRED_FIELDS,
-                                "long-tail row")
-        self.assertEqual(row["id"], 1)
-        self.assertEqual(row["band"], "missing")
-        self.assertFalse(row["in_flight_rescue"])
-        # Wire shape IS the Struct shape — round-trips cleanly.
-        back = msgspec.convert(row, type=LongTailRow)
-        self.assertEqual(back.id, 1)
-
-    def test_transparent_band_via_beets_seam(self):
-        """AE2 at the HTTP boundary: a wanted row whose beets copy
-        classifies Transparent bands ``transparent``. The beets leaf
-        seam is patched to report the release in-library with a
-        lossless detail row."""
-        self.db.seed_request(make_request_row(
-            id=1, status="wanted", mb_release_id="rel-1"))
-
-        beets_db = FakeBeetsDB()
-        # MP3 @ 256 kbps classifies TRANSPARENT in the default rank model
-        # (Opus 128 / MP3 V0 are transparent; see docs/quality-ranks.md).
-        beets_db.set_mbid_detail(
-            "rel-1", {"beets_format": "MP3", "beets_bitrate": 256})
-        with patch("web.server.check_beets_library",
-                   return_value={"rel-1"}), \
-                patch("web.server._beets_db", return_value=beets_db):
-            status, data = self._get("/api/pipeline/long-tail")
-
-        self.assertEqual(status, 200)
-        self.assertEqual(data["results"][0]["band"], "transparent")
-
-    def test_unknown_band_when_in_library_but_unrankable(self):
-        """In-library but no detail / unrankable → ``unknown``, not
-        ``missing``."""
-        self.db.seed_request(make_request_row(
-            id=1, status="wanted", mb_release_id="rel-1"))
-
-        beets_db = FakeBeetsDB()  # no detail row seeded
-        with patch("web.server.check_beets_library",
-                   return_value={"rel-1"}), \
-                patch("web.server._beets_db", return_value=beets_db):
-            status, data = self._get("/api/pipeline/long-tail")
-
-        self.assertEqual(status, 200)
-        self.assertEqual(data["results"][0]["band"], "unknown")
-
-    def test_in_flight_rescue_stamped(self):
-        self.db.seed_request(make_request_row(
-            id=1, status="wanted", mb_release_id="rel-1"))
-        self.db.insert_youtube_running(
-            request_id=1, browse_id="MPREb_z", audio_playlist_id=None,
-            yt_url="https://music.youtube.com/playlist?list=z",
-            expected_track_count=10,
-        )
-        status, data = self._get("/api/pipeline/long-tail")
-        self.assertEqual(status, 200)
-        self.assertTrue(data["results"][0]["in_flight_rescue"])
-
-    def test_band_filter_narrows_result(self):
-        self.db.seed_request(make_request_row(
-            id=1, status="wanted", mb_release_id="rel-1"))
-        self.db.seed_request(make_request_row(
-            id=2, status="wanted", mb_release_id="rel-2"))
-        # No beets → both Missing.
-        status, data = self._get("/api/pipeline/long-tail?band=missing")
-        self.assertEqual(status, 200)
-        self.assertEqual(data["band"], "missing")
-        self.assertEqual({r["id"] for r in data["results"]}, {1, 2})
-        # A band with no members returns an empty cohort, still 200.
-        status, data = self._get("/api/pipeline/long-tail?band=transparent")
-        self.assertEqual(status, 200)
-        self.assertEqual(data["count"], 0)
-
-    def test_empty_cohort_returns_200(self):
-        status, data = self._get("/api/pipeline/long-tail")
-        self.assertEqual(status, 200)
-        self.assertEqual(data["count"], 0)
-        self.assertEqual(data["results"], [])
-
-    def test_single_id_returns_one_banded_row(self):
-        """KTD8: ``?id=`` returns just that request's authoritative band."""
-        from lib.long_tail_service import LongTailRow
-        self.db.seed_request(make_request_row(
-            id=42, status="wanted", mb_release_id="rel-42",
-            artist_name="One", album_title="Row"))
-        status, data = self._get("/api/pipeline/long-tail?id=42")
-        self.assertEqual(status, 200)
-        _assert_required_fields(self, data, {"result", "id"},
-                                "long-tail single-id envelope")
-        self.assertEqual(data["id"], 42)
-        row = msgspec.convert(data["result"], type=LongTailRow)
-        self.assertEqual(row.id, 42)
-        self.assertEqual(row.band, "missing")
-
-    def test_single_id_404_when_not_wanted(self):
-        self.db.seed_request(make_request_row(
-            id=42, status="imported", mb_release_id="rel-42"))
-        status, data = self._get("/api/pipeline/long-tail?id=42")
-        self.assertEqual(status, 404)
-        self.assertEqual(data["id"], 42)
-
-    def test_single_id_400_on_non_int(self):
-        status, data = self._get("/api/pipeline/long-tail?id=not-an-int")
-        self.assertEqual(status, 400)
-        self.assertIn("error", data)
 
 if __name__ == "__main__":
     unittest.main()
