+++ web/routes/pipeline.py
@@ -1,8 +1,14 @@
-"""Pipeline API route handlers, extracted from server.py."""
+"""Pipeline API route handlers, extracted from server.py.
+
+Core pipeline CRUD, dashboard/decisions, import jobs, and beets-distance.
+Search-plan, triage, long-tail, and the /api/_index self-documentation
+machinery were split out (#481 item 3) into sibling modules:
+``web/routes/search_plan.py``, ``web/routes/triage.py``,
+``web/routes/long_tail.py``, ``web/routes/api_index.py``.
+"""
 
 import json
 import logging
-import textwrap
 import urllib.error
 from pathlib import Path
 from typing import Literal
@@ -628,481 +634,6 @@ def get_pipeline_detail(h, params: dict[str, list[str]], req_id_str: str) -> Non
         if tracks is not None:
             result["beets_tracks"] = tracks
     h._json(result)
-
-
-def get_pipeline_search_plan(
-    h, params: dict[str, list[str]], req_id_str: str,
-) -> None:
-    """U6: read-only view of a request's persisted search plan.
-
-    Mirrors ``pipeline-cli search-plan show --json`` so the future
-    dashboard and operators see the same JSON. The U8 stats block is
-    included by default; pass ``stats=0`` to suppress it for a leaner
-    payload (the show endpoint stays a single contract).
-    """
-    from lib.search_plan_inspection import (
-        RequestNotFound,
-        build_inspection_payload,
-    )
-    include_stats = params.get("stats", ["1"])[0] != "0"
-    db = _server()._db()
-    payload = build_inspection_payload(
-        db, int(req_id_str), include_stats=include_stats)
-    if isinstance(payload, RequestNotFound):
-        h._error("Not found", 404)
-        return
-    h._json(payload)
-
-
-def get_pipeline_search_plan_dry_run(
-    h, params: dict[str, list[str]], req_id_str: str,
-) -> None:
-    """U6: ``GET /api/pipeline/<id>/search-plan/dry-run``.
-
-    Read-only generator simulator. Runs ``generate_search_plan``
-    against the current persisted snapshot for ``<id>`` and returns
-    the resulting plan items without writing anything. Counterpart of
-    ``pipeline-cli search-plan dry-run``; both surfaces wrap
-    ``SearchPlanService.dry_run_for_request`` — keep them in sync
-    (see ``CLAUDE.md`` § "CLI ⇄ API surface symmetry").
-
-    Query string:
-      * ``prepend_artist`` — ``1`` to enable (defaults to runtime
-        config's ``album_prepend_artist``).
-
-    Status-code mapping:
-      * 200 — ``RESULT_DRY_RUN_SUCCESS`` or
-        ``RESULT_DRY_RUN_GENERATION_FAILED`` (the latter is a
-        deterministic generator outcome the operator wants to see in
-        the body — the route is informational, not a hard error).
-      * 404 — ``RESULT_REQUEST_NOT_FOUND``.
-    """
-    from lib.config import read_runtime_config
-    from lib.search_plan_service import (
-        RESULT_DRY_RUN_GENERATION_FAILED,
-        RESULT_DRY_RUN_SUCCESS,
-        RESULT_REQUEST_NOT_FOUND,
-        SearchPlanService,
-        dry_run_payload,
-    )
-    try:
-        request_id = int(req_id_str)
-    except (TypeError, ValueError):
-        h._error("Invalid request id")
-        return
-    prepend_artist_raw = params.get("prepend_artist", [None])[0]
-    prepend_artist: bool | None
-    if prepend_artist_raw is None or prepend_artist_raw == "":
-        prepend_artist = None
-    else:
-        prepend_artist = prepend_artist_raw == "1"
-
-    db = _server()._db()
-    cfg = read_runtime_config()
-    svc = SearchPlanService(db, cfg)
-    result = svc.dry_run_for_request(
-        request_id, prepend_artist=prepend_artist,
-    )
-    row = db.get_request(request_id)
-    has_active = False
-    if row is not None:
-        try:
-            active = db.get_active_search_plan(request_id)
-            has_active = active is not None
-        except Exception:  # noqa: BLE001
-            has_active = False
-    payload = dry_run_payload(
-        result,
-        current_generator_id=svc.generator_id,
-        request_row=row,
-        has_active_plan=has_active,
-    )
-    if result.outcome == RESULT_REQUEST_NOT_FOUND:
-        payload["error"] = result.error_message or "Not found"
-        h._json(payload, status=404)
-        return
-    if result.outcome in (
-            RESULT_DRY_RUN_SUCCESS, RESULT_DRY_RUN_GENERATION_FAILED):
-        h._json(payload)
-        return
-    # Defensive fallback for any future outcome string.
-    h._error(f"Unknown dry-run outcome: {result.outcome}", 500)
-
-
-def get_pipeline_search_plan_saturation(
-    h, params: dict[str, list[str]], req_id_str: str,
-) -> None:
-    """U7: ``GET /api/pipeline/<id>/search-plan/saturation``.
-
-    Read-only telemetry aggregator. Returns the saturation rate (rows
-    whose ``final_state`` contains ``LimitReached``) and total
-    ``pre_filter_skip_count`` over the last ``window_days`` (default
-    14) of ``search_log`` rows for ``<id>``. Counterpart of
-    ``pipeline-cli search-plan saturation``; both surfaces wrap
-    ``SearchPlanService.saturation_for_request`` — keep them in sync
-    (see ``CLAUDE.md`` § "CLI ⇄ API surface symmetry").
-
-    Query string:
-      * ``window_days`` — int in
-        ``[SATURATION_WINDOW_MIN_DAYS, SATURATION_WINDOW_MAX_DAYS]``;
-        defaults to ``SATURATION_WINDOW_DEFAULT_DAYS``.
-
-    Status-code mapping:
-      * 200 — ``RESULT_SATURATION_SUCCESS`` (counts may be zero — a
-        valid "found but quiet" state).
-      * 400 — ``window_days`` not parseable as int, or
-        ``RESULT_SATURATION_INPUT_INVALID``.
-      * 404 — ``RESULT_REQUEST_NOT_FOUND`` (request_id does not exist
-        in ``album_requests``; distinct from a real request whose
-        window happens to be empty).
-    """
-    from lib.config import read_runtime_config
-    from lib.search_plan_service import (
-        RESULT_REQUEST_NOT_FOUND,
-        RESULT_SATURATION_INPUT_INVALID,
-        RESULT_SATURATION_SUCCESS,
-        SATURATION_WINDOW_DEFAULT_DAYS,
-        SearchPlanService,
-        saturation_payload,
-    )
-    try:
-        request_id = int(req_id_str)
-    except (TypeError, ValueError):
-        h._error("Invalid request id")
-        return
-    window_raw = params.get("window_days", [None])[0]
-    if window_raw is None or window_raw == "":
-        window_days = SATURATION_WINDOW_DEFAULT_DAYS
-    else:
-        try:
-            window_days = int(window_raw)
-        except (TypeError, ValueError):
-            h._error(
-                f"window_days must be an integer; got {window_raw!r}",
-                status=400,
-            )
-            return
-
-    db = _server()._db()
-    cfg = read_runtime_config()
-    svc = SearchPlanService(db, cfg)
-    result = svc.saturation_for_request(
-        request_id, window_days=window_days,
-    )
-    payload = saturation_payload(result)
-    if result.outcome == RESULT_REQUEST_NOT_FOUND:
-        payload["error"] = result.error_message or "Not found"
-        h._json(payload, status=404)
-        return
-    if result.outcome == RESULT_SATURATION_INPUT_INVALID:
-        payload["error"] = (
-            result.error_message or "Invalid window_days")
-        h._json(payload, status=400)
-        return
-    if result.outcome == RESULT_SATURATION_SUCCESS:
-        h._json(payload)
-        return
-    # Defensive fallback for any future outcome string.
-    h._error(f"Unknown saturation outcome: {result.outcome}", 500)
-
-
-def get_pipeline_search_plan_history(
-    h, params: dict[str, list[str]], req_id_str: str,
-) -> None:
-    """``GET /api/pipeline/<id>/search-plan/history``.
-
-    Cursor-paginated read of one request's ``search_log`` rows. Wraps
-    ``SearchPlanService.history_for_request``; both this route and
-    ``pipeline-cli search-plan history`` go through the same service
-    method so the input bounds and outcome mapping cannot drift.
-
-    Query string:
-      * ``limit`` — int in ``[1, 200]``; defaults to
-        ``HISTORY_PAGE_DEFAULT_LIMIT`` when omitted.
-      * ``before_id`` — int >= 1; the ``next_before_id`` from the
-        previous page. Omit for the first page.
-
-    Status-code mapping:
-      * 200 — ``RESULT_HISTORY_PAGE_SUCCESS``
-      * 400 — query string non-int / ``RESULT_HISTORY_PAGE_INPUT_INVALID``
-      * 404 — ``RESULT_REQUEST_NOT_FOUND``
-    """
-    from lib.config import read_runtime_config
-    from lib.search_plan_service import (
-        HISTORY_PAGE_DEFAULT_LIMIT,
-        RESULT_HISTORY_PAGE_INPUT_INVALID,
-        RESULT_HISTORY_PAGE_SUCCESS,
-        RESULT_REQUEST_NOT_FOUND,
-        SearchPlanService,
-    )
-    try:
-        request_id = int(req_id_str)
-    except (TypeError, ValueError):
-        h._error("Invalid request id")
-        return
-    limit_raw = params.get("limit", [None])[0]
-    if limit_raw is None or limit_raw == "":
-        limit = HISTORY_PAGE_DEFAULT_LIMIT
-    else:
-        try:
-            limit = int(limit_raw)
-        except (TypeError, ValueError):
-            h._error("limit must be an integer")
-            return
-    before_id_raw = params.get("before_id", [None])[0]
-    before_id: int | None
-    if before_id_raw is None or before_id_raw == "":
-        before_id = None
-    else:
-        try:
-            before_id = int(before_id_raw)
-        except (TypeError, ValueError):
-            h._error("before_id must be an integer")
-            return
-
-    db = _server()._db()
-    cfg = read_runtime_config()
-    svc = SearchPlanService(db, cfg)
-    result = svc.history_for_request(
-        request_id, limit=limit, before_id=before_id,
-    )
-    s = _server()
-    if result.outcome == RESULT_HISTORY_PAGE_SUCCESS:
-        # F1: map rows through _serialize_row so datetime fields
-        # (created_at) become ISO strings before json.dumps is called.
-        # F5: pass candidates through msgspec.convert + msgspec.to_builtins
-        # for symmetric wire-boundary strictness (mirrors _build_last_search_payload).
-        serialized_rows: list[dict[str, object]] = []
-        for r in result.rows:
-            sr = s._serialize_row(r)
-            raw_candidates = sr.get("candidates")
-            if raw_candidates is not None:
-                try:
-                    candidates = msgspec.convert(
-                        raw_candidates, type=list[CandidateScore])
-                    sr["candidates"] = [
-                        msgspec.to_builtins(c) for c in candidates]
-                except msgspec.ValidationError as exc:
-                    logger.warning(
-                        "search_log.candidates JSONB failed msgspec validation "
-                        "(request_id=%s, search_log_id=%s): %s",
-                        r.get("request_id"), r.get("id"), exc,
-                    )
-                    sr["candidates"] = None
-            serialized_rows.append(sr)
-        payload: dict[str, object] = {
-            "request_id": result.request_id,
-            "rows": serialized_rows,
-            "next_before_id": result.next_before_id,
-        }
-        h._json(payload)
-        return
-    if result.outcome == RESULT_REQUEST_NOT_FOUND:
-        # F3: match h._error() shape used by neighbor routes (get_pipeline_detail etc.)
-        h._error(result.error_message or "Request not found", 404)
-        return
-    if result.outcome == RESULT_HISTORY_PAGE_INPUT_INVALID:
-        h._error(result.error_message or "Invalid history page request", 400)
-        return
-    # Defensive fallback for any future outcome string.
-    h._error(f"Unknown history outcome: {result.outcome}", 500)
-
-
-class PipelineSearchPlanRegenerateRequest(BaseModel):
-    # ``strict=True`` because Pydantic v2's default lax mode coerces
-    # ``"true"``/``"false"`` strings to bool — the regenerate route's
-    # contract is JSON-bool only and the test pins that.
-    prepend_artist: bool | None = Field(default=None, strict=True)
-
-
-def post_pipeline_search_plan_regenerate(
-    h, body: dict, req_id_str: str,
-) -> None:
-    """U8: ``POST /api/pipeline/<id>/search-plan/regenerate``.
-
-    Wraps ``SearchPlanService.generate_for_request(regenerate=True)``.
-    Allowed for any request status; only ``wanted`` requests are
-    executable, surfaced via ``executable`` in the response so
-    operators can't misread "regenerated" as "now downloading".
-
-    Status-code mapping mirrors the CLI's exit codes:
-      * 200 — ``RESULT_SUCCESS`` or ``RESULT_NOOP_ACTIVE_PLAN_EXISTS``
-      * 404 — ``RESULT_REQUEST_NOT_FOUND``
-      * 422 — ``RESULT_FAILED_DETERMINISTIC`` (sticky, body explains)
-      * 503 — ``RESULT_FAILED_TRANSIENT`` (retryable)
-    """
-    from lib.config import read_runtime_config
-    from lib.search_plan_service import (
-        RESULT_FAILED_DETERMINISTIC,
-        RESULT_FAILED_TRANSIENT,
-        RESULT_NOOP_ACTIVE_PLAN_EXISTS,
-        RESULT_REQUEST_NOT_FOUND,
-        RESULT_SUCCESS,
-        SearchPlanService,
-    )
-    try:
-        request_id = int(req_id_str)
-    except (TypeError, ValueError):
-        h._error("Invalid request id")
-        return
-    req_body = parse_body(h, body or {}, PipelineSearchPlanRegenerateRequest)
-    if req_body is None:
-        return
-    prepend_artist: bool | None = req_body.prepend_artist
-    db = _server()._db()
-    cfg = read_runtime_config()
-    svc = SearchPlanService(db, cfg)
-    result = svc.generate_for_request(
-        request_id, regenerate=True, prepend_artist=prepend_artist,
-    )
-
-    payload: dict[str, object] = {
-        "request_id": request_id,
-        "outcome": result.outcome,
-        "plan_id": result.plan_id,
-        "is_supersede": result.is_supersede,
-        "failure_class": result.failure_class,
-        "error_message": result.error_message,
-    }
-
-    req = db.get_request(request_id)
-    if req is not None:
-        payload["request_status"] = req.get("status")
-        payload["executable"] = (
-            req.get("status") == "wanted"
-            and result.outcome == RESULT_SUCCESS
-        )
-    else:
-        payload["request_status"] = None
-        payload["executable"] = False
-
-    if result.outcome == RESULT_REQUEST_NOT_FOUND:
-        # Symmetric body shape with 422 / 503: clients expect to see
-        # request_id / outcome / plan_id (None) / failure_class /
-        # error_message even on the not-found path.
-        payload["error"] = "Not found"
-        h._json(payload, status=404)
-        return
-    if result.outcome == RESULT_FAILED_DETERMINISTIC:
-        payload["error"] = result.error_message or "Plan generation failed"
-        h._json(payload, status=422)
-        return
-    if result.outcome == RESULT_FAILED_TRANSIENT:
-        payload["error"] = result.error_message or "Plan generation retryable"
-        h._json(payload, status=503)
-        return
-    # RESULT_SUCCESS or RESULT_NOOP_ACTIVE_PLAN_EXISTS.
-    if result.outcome not in (RESULT_SUCCESS, RESULT_NOOP_ACTIVE_PLAN_EXISTS):
-        # Defensive fallback; surface as 500 so we notice unknown shapes.
-        h._error(f"Unknown plan generation outcome: {result.outcome}", 500)
-        return
-    h._json(payload)
-
-
-class PipelineSearchPlanAdvanceRequest(BaseModel):
-    """HTTP body for ``POST /api/pipeline/<id>/search-plan/advance``.
-
-    Exactly one of ``to_ordinal`` / ``to_strategy`` is required. The
-    ``@model_validator`` enforces the XOR — Pydantic checks types but
-    not "exactly one of two".
-    """
-
-    to_ordinal: int | None = None
-    to_strategy: str | None = None
-
-    @model_validator(mode="after")
-    def _exactly_one_target(self) -> "PipelineSearchPlanAdvanceRequest":
-        if (self.to_ordinal is None) == (self.to_strategy is None):
-            raise ValueError(
-                "exactly one of to_ordinal or to_strategy is required"
-            )
-        return self
-
-
-def post_pipeline_search_plan_advance(
-    h, body: dict, req_id_str: str,
-) -> None:
-    """``POST /api/pipeline/<id>/search-plan/advance``.
-
-    Forward-only operator advance of the search-plan cursor. Counterpart
-    of ``pipeline-cli search-plan advance``. Both surfaces wrap
-    ``SearchPlanService.advance_for_request`` — keep them in sync (see
-    ``CLAUDE.md`` § "CLI ⇄ API surface symmetry").
-
-    Body: exactly one of
-      * ``{"to_ordinal": <int>}`` — absolute target ordinal
-      * ``{"to_strategy": <str>}`` — first slot past cursor whose strategy
-        starts with this prefix
-
-    Status-code mapping mirrors the CLI exit codes:
-      * 200 — ``RESULT_ADVANCED``
-      * 400 — body validation failure (missing/extra keys, wrong type)
-      * 404 — ``RESULT_REQUEST_NOT_FOUND``
-      * 409 — ``RESULT_NO_ACTIVE_PLAN`` (request needs ``regenerate`` first)
-      * 422 — ``RESULT_INVALID_TARGET`` (out of range, backward, no slot
-        matches strategy)
-      * 503 — ``RESULT_FAILED_TRANSIENT`` (lock contention)
-    """
-    from lib.config import read_runtime_config
-    from lib.search_plan_service import (
-        RESULT_ADVANCED,
-        RESULT_FAILED_TRANSIENT,
-        RESULT_INVALID_TARGET,
-        RESULT_NO_ACTIVE_PLAN,
-        RESULT_REQUEST_NOT_FOUND,
-        SearchPlanService,
-    )
-    try:
-        request_id = int(req_id_str)
-    except (TypeError, ValueError):
-        h._error("Invalid request id")
-        return
-    req_body = parse_body(h, body or {}, PipelineSearchPlanAdvanceRequest)
-    if req_body is None:
-        return
-
-    db = _server()._db()
-    cfg = read_runtime_config()
-    svc = SearchPlanService(db, cfg)
-    result = svc.advance_for_request(
-        request_id,
-        to_ordinal=req_body.to_ordinal,
-        to_strategy=req_body.to_strategy,
-    )
-    payload: dict[str, object] = {
-        "request_id": result.request_id,
-        "outcome": result.outcome,
-        "plan_id": result.plan_id,
-        "previous_ordinal": result.previous_ordinal,
-        "new_ordinal": result.new_ordinal,
-        "new_strategy": result.new_strategy,
-        "new_query": result.new_query,
-        "error_message": result.error_message,
-    }
-    if result.outcome == RESULT_ADVANCED:
-        h._json(payload)
-        return
-    if result.outcome == RESULT_REQUEST_NOT_FOUND:
-        payload["error"] = result.error_message or "Not found"
-        h._json(payload, status=404)
-        return
-    if result.outcome == RESULT_NO_ACTIVE_PLAN:
-        payload["error"] = (
-            result.error_message or "No active plan; regenerate first")
-        h._json(payload, status=409)
-        return
-    if result.outcome == RESULT_INVALID_TARGET:
-        payload["error"] = (
-            result.error_message or "Invalid advance target")
-        h._json(payload, status=422)
-        return
-    if result.outcome == RESULT_FAILED_TRANSIENT:
-        payload["error"] = (
-            result.error_message or "Plan lock contention; retry")
-        h._json(payload, status=503)
-        return
-    # Defensive: any unknown outcome string is a bug.
-    h._error(f"Unknown advance outcome: {result.outcome}", 500)
 
 
 def post_pipeline_resolve_rg(h, body: dict, req_id_str: str) -> None:
@@ -2484,363 +2015,7 @@ def get_beets_distance(
     h._json(payload, status=status)
 
 
-# --- U17: /api/triage HTTP endpoints --------------------------------------
-#
-# Two HTTP routes wrap the U15 triage service (``lib.triage_service``):
-#
-#   * ``GET /api/triage/<id>`` — per-request composition. Mirrors
-#     ``pipeline-cli triage show <id>`` (U16). Outcome → status:
-#       - 200: ``TriageResult`` payload (msgspec.to_builtins).
-#       - 400: non-int request id (h._error default).
-#       - 404: request_id has no album_requests row.
-#
-#   * ``GET /api/triage/list`` — cohort listing. Mirrors
-#     ``pipeline-cli triage list --filter=<spec>`` (U16). Outcome →
-#     status:
-#       - 200: ``{results, next_after, page_size, filter}`` payload.
-#       - 400: ``InvalidFilterError`` or non-int ``limit``/``after``.
-#
-# Both surfaces route through the same service entrypoints
-# (``compose_triage_for_request`` / ``list_triage``) so the CLI ⇄ API
-# symmetry rule holds — see ``CLAUDE.md`` § "CLI ⇄ API surface symmetry".
-
-# Filter forms surfaced in the 400 body — single source of truth lives
-# in ``lib.triage_service.VALID_FILTER_FORMS`` so the CLI and the HTTP
-# 400 envelope advertise the same vocabulary.
-from lib.triage_service import VALID_FILTER_FORMS as _TRIAGE_VALID_FILTER_FORMS_API  # noqa: E402
-
-# Page-size bounds for ``GET /api/triage/list`` — re-exports of the
-# single-source-of-truth constants on ``lib.triage_service`` so the CLI
-# and API enforce the same ranges. Mirrors the convention established by
-# ``get_pipeline_search_plan_history`` (1..200): a hard upper bound
-# prevents an unbounded scan; the lower bound rules out the nonsense
-# ``limit=0`` request shape.
-from lib.triage_service import (  # noqa: E402
-    DEFAULT_TRIAGE_PAGE_SIZE as _TRIAGE_LIST_DEFAULT_LIMIT,
-    TRIAGE_AFTER_MIN as _TRIAGE_LIST_MIN_AFTER,
-    TRIAGE_LIMIT_MAX as _TRIAGE_LIST_MAX_LIMIT,
-    TRIAGE_LIMIT_MIN as _TRIAGE_LIST_MIN_LIMIT,
-)
-
-
-def get_triage_for_request(
-    h, params: dict[str, list[str]], req_id_str: str,
-) -> None:
-    """U17: ``GET /api/triage/<id>``.
-
-    Compose the per-request triage payload via
-    ``lib.triage_service.compose_triage_for_request``. The response
-    body is ``msgspec.to_builtins(TriageResult)`` — the JSON shape on
-    the wire IS the Struct shape verbatim, which is what makes
-    ``msgspec.convert(payload, type=TriageResult)`` round-trip on the
-    consumer side (frontend or CLI parity tests).
-
-    Status-code mapping (mirrors ``cmd_triage_show``'s exit codes):
-      * 200 — composition success.
-      * 404 — ``compose_triage_for_request`` returned ``None`` (no row).
-
-    The route's regex (``r"^/api/triage/(\\d+)$"``) requires a digit-only
-    path segment, so ``req_id_str`` is always coercible — non-digit
-    paths never match this pattern in the first place and fall through
-    to the catch-all 404 in ``web/server.py``.
-    """
-    from lib.triage_service import compose_triage_for_request
-
-    request_id = int(req_id_str)
-    db = _server()._db()
-    result = compose_triage_for_request(request_id, db)
-    if result is None:
-        h._json(
-            {"error": "Not found", "request_id": request_id},
-            status=404,
-        )
-        return
-
-    payload = msgspec.to_builtins(result)
-    h._json(payload)
-
-
-def get_triage_list(
-    h, params: dict[str, list[str]],
-) -> None:
-    """U17: ``GET /api/triage/list``.
-
-    Cohort-filtered triage listing. Query string:
-      * ``filter`` — filter spec (default ``"all"``). Forms documented
-        in ``_TRIAGE_VALID_FILTER_FORMS_API`` and ``lib.triage_service
-        .parse_filter``.
-      * ``limit`` — int in ``[_TRIAGE_LIST_MIN_LIMIT,
-        _TRIAGE_LIST_MAX_LIMIT]``; defaults to
-        ``_TRIAGE_LIST_DEFAULT_LIMIT``.
-      * ``after`` — int >= 1; the ``next_after`` cursor from the
-        previous page. Omit for the first page.
-
-    Response shape (success):
-        ``{"results": [...], "next_after": <int|null>,
-           "page_size": <int>, "filter": <spec str>}``
-
-    ``next_after`` is ``None`` when ``len(results) < page_size`` (the
-    page exhausts the cohort); otherwise the last request id so
-    operators can keep paging.
-
-    Status-code mapping (mirrors ``cmd_triage_list``'s exit codes):
-      * 200 — success (empty results list is a valid cohort state).
-      * 400 — ``InvalidFilterError`` (parser rejects the spec) OR
-              non-int ``limit`` / ``after`` / out-of-range ``limit``.
-    """
-    from lib.triage_service import InvalidFilterError, list_triage
-
-    filter_spec = params.get("filter", ["all"])[0]
-    if filter_spec == "":
-        filter_spec = "all"
-
-    limit_raw = params.get("limit", [None])[0]
-    if limit_raw is None or limit_raw == "":
-        limit = _TRIAGE_LIST_DEFAULT_LIMIT
-    else:
-        try:
-            limit = int(limit_raw)
-        except (TypeError, ValueError):
-            h._error("limit must be an integer")
-            return
-    if not (_TRIAGE_LIST_MIN_LIMIT <= limit <= _TRIAGE_LIST_MAX_LIMIT):
-        h._error(
-            f"limit must be in [{_TRIAGE_LIST_MIN_LIMIT}, "
-            f"{_TRIAGE_LIST_MAX_LIMIT}]",
-            status=400,
-        )
-        return
-
-    after_raw = params.get("after", [None])[0]
-    after: int | None
-    if after_raw is None or after_raw == "":
-        after = None
-    else:
-        try:
-            after = int(after_raw)
-        except (TypeError, ValueError):
-            h._error("after must be an integer")
-            return
-        if after < _TRIAGE_LIST_MIN_AFTER:
-            h._error(
-                f"after must be >= {_TRIAGE_LIST_MIN_AFTER}", status=400,
-            )
-            return
-
-    db = _server()._db()
-    try:
-        results = list_triage(
-            filter_spec, db, page_size=limit, after_request_id=after,
-        )
-    except InvalidFilterError as exc:
-        # Pull the parameter-vocab arrays so API-only operators can
-        # self-correct from the response body alone (e.g. on
-        # ``unfindable:<bad_cat>``, the operator sees the four valid
-        # categories without needing to consult --help).
-        from lib.triage_service import (
-            VALID_DATA_QUALITY_FIELD_NAMES,
-            VALID_UNFINDABLE_CATEGORIES,
-        )
-        h._json(
-            {
-                "error": str(exc),
-                "valid_filters": list(_TRIAGE_VALID_FILTER_FORMS_API),
-                "valid_unfindable_categories": sorted(
-                    VALID_UNFINDABLE_CATEGORIES
-                ),
-                "valid_data_quality_fields": sorted(
-                    VALID_DATA_QUALITY_FIELD_NAMES
-                ),
-            },
-            status=400,
-        )
-        return
-
-    next_after: int | None = None
-    if len(results) >= limit and results:
-        next_after = results[-1].request_meta.id
-
-    payload: dict[str, object] = {
-        "results": msgspec.to_builtins(results),
-        "next_after": next_after,
-        "page_size": limit,
-        "filter": filter_spec,
-    }
-    h._json(payload)
-
-
-def get_pipeline_long_tail(h, params: dict[str, list[str]]) -> None:
-    """U1: ``GET /api/pipeline/long-tail`` — banded ``wanted`` worklist.
-
-    Returns every ``wanted`` request pre-banded by on-disk quality
-    (``missing`` / a lowercase ``QualityRank`` band / ``unknown``) and
-    stamped with ``in_flight_rescue``. Wraps
-    ``lib.long_tail_service.list_long_tail`` — the SAME service method
-    ``pipeline-cli long-tail`` wraps (CLI ⇄ API symmetry, R16).
-
-    Query string:
-      * ``band`` — optional single-band filter (``missing`` / a rank /
-        ``unknown``). The UI fetches unfiltered and filters client-side;
-        this backs scripted / CLI-parity callers.
-      * ``id`` — optional single request id. Returns just that request's
-        authoritative band + flags (KTD8 — the post-action single-row
-        refetch). 404 when the id doesn't exist OR is no longer
-        ``wanted`` (an imported / replaced row is correctly absent from
-        the worklist). Mutually exclusive with ``band`` (``band`` is
-        ignored when ``id`` is present — a single row is already its own
-        cohort).
-
-    Response shape (cohort success):
-        ``{"results": [...], "band": <str|null>, "count": <int>}``
-    Response shape (single-id success):
-        ``{"result": <row>, "id": <int>}``
-
-    Each result is a ``LongTailRow`` serialized via
-    ``msgspec.to_builtins`` so ``msgspec.convert(row, type=LongTailRow)``
-    round-trips on the consumer side. Datetime / UUID columns are pinned
-    out of the row by ``_LONG_TAIL_SELECT``'s projection, but the row is
-    still routed through ``_serialize_row`` to guard the datetime-500
-    class against any future column add.
-
-    Status-code mapping:
-      * 200 — success (empty cohort is a valid state).
-      * 400 — non-int ``id``.
-      * 404 — ``id`` not found / not ``wanted``.
-    """
-    from lib.long_tail_service import band_one_long_tail, list_long_tail
-    from web.routes._overlay import band_release_ids
-
-    s = _server()
-
-    id_raw = params.get("id", [None])[0]
-    if id_raw is not None and id_raw != "":
-        try:
-            request_id = int(id_raw)
-        except (TypeError, ValueError):
-            h._error("id must be an integer")
-            return
-        row = band_one_long_tail(s._db(), band_release_ids, request_id)
-        if row is None:
-            h._json(
-                {"error": "Not found", "id": request_id},
-                status=404,
-            )
-            return
-        serialized = s._serialize_row(msgspec.to_builtins(row))
-        h._json({"result": serialized, "id": request_id})
-        return
-
-    band = params.get("band", [None])[0]
-    if band == "":
-        band = None
-
-    result = list_long_tail(s._db(), band_release_ids, band=band)
-
-    # Route the serialized rows through _serialize_row to convert any
-    # datetime / UUID columns to JSON-safe values (datetime-500 guard).
-    rows = [s._serialize_row(r) for r in msgspec.to_builtins(result.rows)]
-    h._json({
-        "results": rows,
-        "band": result.band_filter,
-        "count": len(rows),
-    })
-
-
-# --- U18 step 2: /api/_index — self-documenting API surface ----------------
-#
-# Walks ``web.server.Handler``'s merged dispatch tables and emits one row per
-# registered route: path/pattern, method, description, and the Pydantic
-# ``*Request`` model name extracted from the handler's body. The Pydantic
-# field comes from ``inspect.getsource`` + an AST walk for the
-# ``parse_body(h, body, SomeRequest)`` call — see
-# ``code-quality.md`` § "HTTP request bodies — use pydantic.BaseModel".
-#
-# Frontends and the CLI's ``routes`` command both consume this to build
-# self-documenting indexes — keep the response shape stable.
-
-def _extract_request_model(fn: object) -> str | None:
-    """Pull the Pydantic ``*Request`` model name from a POST handler.
-
-    Walks the handler's AST and returns the class name of the first
-    ``parse_body(h, body, X)`` call (third positional argument). Returns
-    ``None`` if the handler doesn't use ``parse_body`` or if the source
-    is unavailable (e.g. .pyc-only deploys).
-
-    Uses the same AST-walk pattern as ``tests/test_pydantic_route_audit.py
-    ::_handler_uses_parse_body`` — no regex brittleness on non-canonical
-    arg shapes.
-    """
-    if not callable(fn):
-        return None
-    import ast
-    import inspect
-    try:
-        source = inspect.getsource(fn)  # type: ignore[arg-type]
-    except (OSError, TypeError):
-        return None
-    try:
-        tree = ast.parse(textwrap.dedent(source))
-    except SyntaxError:
-        return None
-    for node in ast.walk(tree):
-        if not isinstance(node, ast.Call):
-            continue
-        target = node.func
-        if isinstance(target, ast.Name) and target.id == "parse_body":
-            pass
-        elif isinstance(target, ast.Attribute) and target.attr == "parse_body":
-            pass
-        else:
-            continue
-        # Third positional arg is the Pydantic model class.
-        if len(node.args) < 3:
-            continue
-        cls_arg = node.args[2]
-        if isinstance(cls_arg, ast.Name):
-            return cls_arg.id
-        if isinstance(cls_arg, ast.Attribute):
-            return cls_arg.attr
-    return None
-
-
-def get_api_index(h, params: dict[str, list[str]]) -> None:
-    """``GET /api/_index`` — self-documenting API surface.
-
-    Returns a list of ``{method, path, description, request_model}`` rows
-    sorted by ``(method, path)``. ``path`` is the registered string for
-    static routes and the regex pattern string for pattern routes.
-    ``request_model`` is the Pydantic model name for POST handlers that
-    use ``parse_body``; null otherwise.
-
-    #496: reads directly from the merged ``RouteRegistration`` list
-    (``srv.ALL_ROUTES``) rather than four separate dispatch/description
-    dicts — one source of truth for path, method, and description.
-    """
-    from web import server as srv
-
-    entries: list[dict[str, object]] = [
-        {
-            "method": r.method,
-            "path": r.path,
-            "description": r.description,
-            "request_model": (
-                _extract_request_model(r.handler) if r.method == "POST" else None
-            ),
-        }
-        for r in srv.ALL_ROUTES
-    ]
-
-    entries.sort(key=lambda e: (str(e["method"]), str(e["path"])))
-    h._json(entries)
-
-
 ROUTES: list[RouteRegistration] = [
-    route(
-        "GET", "/api/_index", get_api_index,
-        "Self-documenting API surface — every route's path, method, "
-        "description, and Pydantic request model.",
-        classified=True,
-    ),
     route(
         "GET", "/api/pipeline/log", get_pipeline_log,
         "Recent download_log rows with per-row classification badges + "
@@ -2914,29 +2089,6 @@ ROUTES: list[RouteRegistration] = [
         "(Replace-button enable set).",
         classified=True,
     ),
-    route(
-        "GET", "/api/pipeline/long-tail", get_pipeline_long_tail,
-        # U1: Long-Tail Triage Console worklist read. Wraps
-        # ``lib.long_tail_service.list_long_tail`` — same service as
-        # ``pipeline-cli long-tail`` per CLI ⇄ API symmetry.
-        "Long-tail worklist — the full wanted cohort pre-banded by "
-        "on-disk quality (missing / QualityRank band / unknown) and "
-        "stamped with in_flight_rescue. Optional ?band= filter.",
-        classified=True,
-    ),
-    route(
-        "GET", "/api/triage/list", get_triage_list,
-        # U17: /api/triage HTTP endpoints. Per-request composition and
-        # cohort listing both wrap ``lib.triage_service`` (U15) — same
-        # service as ``pipeline-cli triage`` (U16) per CLI ⇄ API symmetry.
-        "Cohort triage listing — filter by unfindable category, "
-        "field-quality field/status/reason, or search-not-converting "
-        "state. ``data_quality:status=<status>`` filters on the "
-        "resolver-status column (e.g. unresolved_4xx_client); "
-        "``data_quality:reason=<code>`` filters on the reason_code "
-        "column (e.g. http_400).",
-        classified=True,
-    ),
     # /api/beets-distance/<download_log_id>/<mbid> — real beets distance
     # for one (download_log_id, mbid) pair. See get_beets_distance above.
     pattern_route(
@@ -2953,32 +2105,6 @@ ROUTES: list[RouteRegistration] = [
         classified=True,
     ),
     pattern_route(
-        "GET", r"^/api/pipeline/(\d+)/search-plan$", get_pipeline_search_plan,
-        "Read-only view of a request's persisted search plan (cursor, "
-        "items, provenance, per-slot stats).",
-        classified=True,
-    ),
-    pattern_route(
-        "GET", r"^/api/pipeline/(\d+)/search-plan/dry-run$",
-        get_pipeline_search_plan_dry_run,
-        "Generator simulator — runs generate_search_plan against the "
-        "current snapshot without writing.",
-        classified=True,
-    ),
-    pattern_route(
-        "GET", r"^/api/pipeline/(\d+)/search-plan/saturation$",
-        get_pipeline_search_plan_saturation,
-        "Saturation rate + pre-filter skip total over a recent search_log "
-        "window for this request.",
-        classified=True,
-    ),
-    pattern_route(
-        "GET", r"^/api/pipeline/(\d+)/search-plan/history$",
-        get_pipeline_search_plan_history,
-        "Cursor-paginated read of one request's search_log rows.",
-        classified=True,
-    ),
-    pattern_route(
         "GET", r"^/api/pipeline/requests-by-rg/([a-f0-9-]{36})$",
         get_pipeline_requests_by_rg,
         "Non-replaced album_requests rows sharing the given release "
@@ -2988,12 +2114,6 @@ ROUTES: list[RouteRegistration] = [
     pattern_route(
         "GET", r"^/api/import-jobs/(\d+)$", get_import_job,
         "Single import-job detail by job id.",
-        classified=True,
-    ),
-    pattern_route(
-        "GET", r"^/api/triage/(\d+)$", get_triage_for_request,
-        "Per-request triage composition — unfindable categorisation, "
-        "field-resolution telemetry, search-log forensics.",
         classified=True,
     ),
     route(
@@ -3037,19 +2157,6 @@ ROUTES: list[RouteRegistration] = [
         "POST", "/api/pipeline/delete", post_pipeline_delete,
         "Delete a pipeline request (blocked when a superseding "
         "request exists).",
-        classified=True,
-    ),
-    pattern_route(
-        "POST", r"^/api/pipeline/(\d+)/search-plan/regenerate$",
-        post_pipeline_search_plan_regenerate,
-        "Regenerate the search plan for a request.",
-        classified=True,
-    ),
-    pattern_route(
-        "POST", r"^/api/pipeline/(\d+)/search-plan/advance$",
-        post_pipeline_search_plan_advance,
-        "Forward-only operator advance of the search-plan cursor "
-        "(by ordinal or strategy prefix).",
         classified=True,
     ),
     pattern_route(
